split main
