+++ generator/__main__.py
@@ -1,21 +1,14 @@
+# sourcery no-metrics
 import argparse
 import os
 import shutil
 
-from jinja2 import Environment, FileSystemLoader
 from tqdm import tqdm
 
 from generator.client import Client
-
-env = Environment(
-    loader=FileSystemLoader("templates"), trim_blocks=True, lstrip_blocks=True
-)
-
-courses_template = env.get_template("courses.jinja2")
-lesson_template = env.get_template("lesson.jinja2")
-material_template = env.get_template("material.jinja2")
-task_template = env.get_template("task.jinja2")
-lessons_template = env.get_template("lessons.jinja2")
+from generator.page_savers import save_lesson_page, save_task_page, \
+    save_material_page, save_courses_page, lessons_template
+from generator.stuff import sections_types
 
 parser = argparse.ArgumentParser(description="yandex lyceum docs generator")
 parser.add_argument("--login", type=str, required=True)
@@ -37,112 +30,6 @@ if os.path.exists("docs"):
 os.makedirs(os.path.join("docs", "courses"))
 shutil.copytree(os.path.join("templates", "css"), os.path.join("docs", "css"))
 shutil.copytree(os.path.join("templates", "js"), os.path.join("docs", "js"))
-
-sections_types = {
-    "classwork": "Классная работа",
-    "homework": "Домашняя работа",
-    "additional": "Дополнительные задачи",
-    "individual-work": "Самостоятельная работа",
-    "control-work": "Контрольная работа",
-}
-
-
-def save_courses_page(courses):
-    with open(os.path.join("docs", "index.html"), "w", encoding="utf-8") as fh:
-        fh.write(courses_template.render(courses=courses))
-
-
-def save_lesson_page(
-    materials, course_id, lesson_id, lesson_tasks_ids, lesson_title, lesson_description
-):
-    lesson_path = os.path.join(
-        "docs", "courses", str(course_id), "lessons", str(lesson_id)
-    )
-    os.mkdir(lesson_path)
-    with open(os.path.join(lesson_path, "index.html"), "w", encoding="utf-8") as fh:
-        fh.write(
-            lesson_template.render(
-                materials=materials,
-                task_groups=lesson_tasks_ids,
-                lesson_titletle=lesson_title,
-                lesson_description=lesson_description,
-            )
-        )
-
-
-def save_material_page(content, short_title, title, course_id, lesson_id, material_id):
-    material_path = os.path.join(
-        "docs",
-        "courses",
-        str(course_id),
-        "lessons",
-        str(lesson_id),
-        "materials",
-        str(material_id),
-    )
-    os.mkdir(material_path)
-    with open(os.path.join(material_path, "index.html"), "w", encoding="utf-8") as fh:
-        fh.write(
-            material_template.render(
-                content=content or "", short_title=short_title, title=title
-            )
-        )
-
-
-def save_task_page(
-    lesson_id,
-    task_id,
-    title,
-    lesson_short_title,
-    task_type,
-    task_title,
-    task_description,
-    lesson_tasks_ids,
-    solution_code,
-    solution_url
-):
-    task_path = os.path.join(
-        "docs",
-        "courses",
-        str(course_id),
-        "lessons",
-        str(lesson_id),
-        "tasks",
-        str(task_id),
-    )
-    os.mkdir(task_path)
-
-    task_ids = []
-    for task_group in lesson_tasks_ids:
-        task_ids.extend(task["id"] for task in task_group["tasks"])
-
-    task_index = task_ids.index(task_id)
-    if task_index == 0:
-        previous_task_id = None
-    else:
-        previous_task_id = task_ids[task_ids.index(task_id) - 1]
-    if task_index == len(task_ids) - 1:
-        next_task_id = None
-    else:
-        next_task_id = task_ids[task_ids.index(task_id) + 1]
-
-    with open(os.path.join(task_path, "index.html"), "w", encoding="utf-8") as fh:
-        fh.write(
-            task_template.render(
-                title=title,
-                lesson_short_title=lesson_short_title,
-                previous_task_id=previous_task_id,
-                next_task_id=next_task_id,
-                task_type=task_type,
-                task_type_title=sections_types[task_type],
-                task_title=task_title,
-                task_description=task_description,
-                task_groups=lesson_tasks_ids,
-                task_id=task_id,
-                solution_code=solution_code,
-                solution_url=solution_url,
-            )
-        )
 
 
 print()
@@ -233,6 +120,7 @@ for course in courses:
                     solution_url = None
                 save_task_page(
                     lesson_id=lesson["id"],
+                    course_id=course_id,
                     task_id=task["id"],
                     title=lesson["title"],
                     lesson_short_title=lesson["shortTitle"],
