+++ generator/stuff.py
@@ -1,0 +1,7 @@
+sections_types = {
+    "classwork": "Классная работа",
+    "homework": "Домашняя работа",
+    "additional": "Дополнительные задачи",
+    "individual-work": "Самостоятельная работа",
+    "control-work": "Контрольная работа",
+}
